all methods work good exept leaveTrip

+++ src/main/java/com/example/tripsters/service/impl/TripServiceImpl.java
@@ -4,6 +4,7 @@ import com.example.tripsters.dto.trip.CreateTripRequestDto;
 import com.example.tripsters.dto.trip.TripResponseDto;
 import com.example.tripsters.dto.trip.UpdateTripRequestDto;
 import com.example.tripsters.exception.EntityNotFoundException;
+import com.example.tripsters.exception.UnauthorizedException;
 import com.example.tripsters.mapper.TripMapper;
 import com.example.tripsters.model.Chat;
 import com.example.tripsters.model.Map;
@@ -36,19 +37,23 @@ public class TripServiceImpl implements TripService {
                 .orElseThrow(() -> new EntityNotFoundException("Authenticated user not found"));
     }
 
+    private void checkOwnership(Trip trip, User user) {
+        if (!trip.getOwnerId().equals(user.getId())) {
+            throw new UnauthorizedException("You are not the owner of this trip");
+        }
+    }
+
     @Override
     public TripResponseDto createTrip(CreateTripRequestDto requestDto) {
-        Trip trip = new Trip();
+        Trip trip = tripMapper.toModel(requestDto);
         User authenticatedUser = getAuthenticatedUser();
         Chat tripChat = new Chat();
         Map tripMap = new Map();
 
-        trip.setDestination(requestDto.getDestination());
-        trip.setStartDate(LocalDateTime.parse(requestDto.getStartDate()));
-        trip.setEndDate(LocalDateTime.parse(requestDto.getEndDate()));
         trip.setCreatedAt(LocalDateTime.now());
         trip.setChat(tripChat);
         trip.setMap(tripMap);
+        trip.setOwnerId(authenticatedUser.getId());
 
         Set<User> listOfUsersInTrip = trip.getUsers();
         listOfUsersInTrip.add(authenticatedUser);
@@ -60,12 +65,16 @@ public class TripServiceImpl implements TripService {
 
     @Override
     public TripResponseDto updateTrip(UpdateTripRequestDto requestDto) {
+        User authenticatedUser = getAuthenticatedUser();
         Trip trip = tripRepository.findById(requestDto.getId())
                 .orElseThrow(() -> new EntityNotFoundException("Trip not found with id: " + requestDto.getId()));
 
+        checkOwnership(trip, authenticatedUser);
         trip.setDestination(requestDto.getDestination());
         trip.setStartDate(LocalDateTime.parse(requestDto.getStartDate()));
         trip.setEndDate(LocalDateTime.parse(requestDto.getEndDate()));
+        trip.setStartAdress(requestDto.getStartAdress());
+        trip.setFinishAdress(requestDto.getFinishAdress());
 
         tripRepository.save(trip);
 
@@ -74,10 +83,16 @@ public class TripServiceImpl implements TripService {
 
     @Override
     public TripResponseDto addUserToTrip(Long tripId, Long userId) {
+        User authenticatedUser = getAuthenticatedUser();
         Trip trip = tripRepository.findById(tripId)
                 .orElseThrow(() -> new EntityNotFoundException("Trip not found with id: " + tripId));
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + userId));
+
+        checkOwnership(trip, authenticatedUser);
+        if (userId.equals(authenticatedUser.getId())) {
+            throw new IllegalArgumentException("You are already in trip");
+        }
 
         Set<User> listOfUserInTrip = trip.getUsers();
         listOfUserInTrip.add(user);
@@ -103,19 +118,25 @@ public class TripServiceImpl implements TripService {
                 .orElseThrow(() -> new EntityNotFoundException("Trip not found with id: " + tripId));
 
         Set<User> listOfUsersInTrip = trip.getUsers();
-        if (!listOfUsersInTrip.remove(authenticatedUser)) {
-            throw new EntityNotFoundException("User not found in the trip with id: " + tripId);
+        if (!trip.getOwnerId().equals(authenticatedUser.getId())) {
+            if (!listOfUsersInTrip.remove(authenticatedUser)) {
+                throw new EntityNotFoundException("User not found in the trip with id: " + tripId);
+            }
+            tripRepository.save(trip);
+        } else {
+            throw new UnauthorizedException("Owner cannot leave the trip");
         }
-
-        tripRepository.save(trip);
 
         return tripMapper.toDto(trip);
     }
 
     @Override
     public void deleteTripById(Long tripId) {
+        User authenticatedUser = getAuthenticatedUser();
         Trip trip = tripRepository.findById(tripId)
                 .orElseThrow(() -> new EntityNotFoundException("Trip not found with id: " + tripId));
+
+        checkOwnership(trip, authenticatedUser);
 
         tripRepository.delete(trip);
     }
